locomotion_simulation_z_800/requirements.txt
absl-py==0.11.0
appdirs==1.4.3
attrs==20.2.0
CacheControl==0.12.6
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
cloudpickle==1.6.0
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
#Cython==0.29.21
decorator==4.4.2
distlib==0.3.1
distro==1.4.0
filelock==3.0.12
filterpy==1.4.5
future==0.18.2
gym==0.17.3
html5lib==1.0.1
identify==1.5.9
idna==2.8
importlab==0.5.1
inputs==0.5
ipaddr==2.2.0
kiwisolver==1.3.1
llvmlite==0.34.0
lockfile==0.12.2
matplotlib==3.3.2
msgpack==0.6.2
networkx==2.5
ninja==1.10.0.post2
nodeenv==1.5.0
numba==0.51.2
numpy==1.19.4
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
pre-commit==2.8.2
progress==1.5
pybullet==3.0.6
pyglet==1.5.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
PyYAML==5.3.1
quadprog==0.1.7
requests==2.22.0
retrying==1.3.3
scipy==1.5.3
six==1.14.0
toml==0.10.2
tqdm==4.51.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.1.0
webencodings==0.5.1
